Port cache classes: `SecretCache` to Task SDK (#49779)
