+++ airflow-core/newsfragments/49779.significant.rst
@@ -1,0 +1,20 @@
+SecretCache class has been moved to ``airflow.sdk.execution_time.cache`` from ``airflow.secrets.cache``
+
+* Types of change
+
+  * [ ] Dag changes
+  * [ ] Config changes
+  * [ ] API changes
+  * [ ] CLI changes
+  * [ ] Behaviour changes
+  * [ ] Plugin changes
+  * [ ] Dependency changes
+  * [x] Code interface changes
+
+* Migration rules needed
+
+  * ruff
+
+    * AIR301
+
+      * [ ] ``airflow.secrets.cache.SecretCache`` → ``airflow.sdk.execution_time.cache.SecretCache``
