+++ airflow-core/tests/unit/secrets/test_cache.py
@@ -21,7 +21,7 @@ import multiprocessing
 
 import pytest
 
-from airflow.secrets.cache import SecretCache
+from airflow.sdk import SecretCache
 
 from tests_common.test_utils.config import conf_vars
